+++ example/lib/main.dart
@@ -23,6 +23,11 @@ class _MyAppState extends State<MyApp> {
     initPushBots();
     registerForNotification();
 
+    Pushbots.listenForNotificationReceive().stream.listen((onData) {
+      print("MAIN, received: " + onData.toString());
+    });
+
+
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.