+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.0.3
+
+* Get notifier working i.e:- 
+
+```dart
+ Pushbots.listenForNotificationReceive().stream.listen((onData) {
+      print("MAIN, received: " + onData.toString());
+    });
+```
+## 0.0.2
+
+* Update Readme.md
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release which support #android
+